feat: show user achievements progress
